timely-util: extract co_reduce2 helpers and tests into submodules
Pure code motion preparing the driver/tactic split. No behavior change.
